// src/app/interfaces/common/pre-shop.interface.ts
export interface PreShop {
  email?: any;
  _id?: string;
  name?: string;
  port?: number;
  slug?: string;
  websiteName?: string;
  paymentRefId?: string;
  paymentMethod?: string;
  paymentStatus?: string;
  amount?: string;
  phoneNo?: string;
  buildStatus?: string;
  images?: any;
  domain?: string;
  subDomain?: string;
  category?: string;
  subCategory?: string;
  theme?: any;
  dateString?: string;
  owner?: any;
  users?: any;
  createdAt?: string;
  updatedAt?: string;
  select: boolean;
}